backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
djoser==2.3.1
psycopg2-binary==2.9.9
Pillow==10.3.0
django-cleanup==8.1.0
django-filter==23.2
gunicorn==21.2.0
fpdf2==2.8.3
uharfbuzz==0.41.0
drf_extra_fields==3.7.0
django-unfold==0.49.1